Move max_poolXd functions to nn.py file. (#617)
So all of max_poolXd functions will be in the same file.
